Add files via upload
The first 2 images are from a project where I wanted to recreate the weapon that John Wick uses in the film franchise's first movie that could potentially be used for an FPS-style game. The second set of images is an old-fashioned saloon that could be used as an area for a western cowboy-style game.